+++ src/src/transports/tf_maildir.h
@@ -1,4 +1,4 @@
-/* $Cambridge: exim/src/src/transports/tf_maildir.h,v 1.2 2005/01/04 10:00:45 ph10 Exp $ */
+/* $Cambridge: exim/src/src/transports/tf_maildir.h,v 1.3 2005/06/07 15:20:56 ph10 Exp $ */
 
 /*************************************************
 *     Exim - an Internet mail transport agent    *
@@ -10,12 +10,12 @@
 /* Header file for the functions that are used to support the use of
 maildirsize files for quota handling in maildir directories. */
 
-extern int  maildir_compute_size(uschar *, int *, time_t *, const pcre *,
-              const pcre *, BOOL);
-extern BOOL maildir_ensure_directories(uschar *, address_item *, BOOL, int);
-extern int  maildir_ensure_sizefile(uschar *,
-              appendfile_transport_options_block *, const pcre *, const pcre *,
-              int *, int *);
-extern void maildir_record_length(int, int);
+extern off_t  maildir_compute_size(uschar *, int *, time_t *, const pcre *,
+                const pcre *, BOOL);
+extern BOOL   maildir_ensure_directories(uschar *, address_item *, BOOL, int);
+extern int    maildir_ensure_sizefile(uschar *,
+                appendfile_transport_options_block *, const pcre *,
+                const pcre *, off_t *, int *);
+extern void   maildir_record_length(int, int);
 
 /* End of tf_maildir.h */
